Revert "Revert "feat(api): Project key creation rate limit params (#15077)" (#15357)"
This reverts commit 9cd81e1493f79af1e7b134cc0c6f65db92d4722f.
